standardize cl arguments
