+++ package.json
@@ -29,6 +29,7 @@
     "test": "jest --silent"
   },
   "dependencies": {
+    "camelcase": "^5.0.0",
     "chalk": "^2.4.1",
     "ejs": "^2.6.1",
     "fs-extra": "^6.0.1",
